requirements.txt
nose==1.3.7
PyYAML==3.11
requests==2.10.0
xmltodict==0.10.1
